gitlab_extractor/requirements.txt
fuzzywuzzy
python-Levenshtein
GitPython
PyYAML
Requests
matplotlib
pandas
seaborn
numpy
kneed
notebook